moved components into their own files
