+++ src/Components/Blog/Blog.tsx
@@ -1,11 +1,14 @@
 import * as React from 'react';
 import createWithAuth from '../../firebase/createWithAuth';
 import { WrappedComponentProps } from 'react-with-firebase-auth';
-import BlogPostRepository, { BlogPost, IBlogPostWithKey } from '../../firebase/BlogPost';
+import BlogPostRepository, { IBlogPostWithKey, BlogPost } from '../../firebase/BlogPost';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { ShowLoading } from '../Shared/ShowLoading';
-import { Box, Button, TextArea, Grommet, Markdown } from 'grommet';
-import { useHistory, Route, Switch, useRouteMatch, useParams } from 'react-router-dom';
+import { Box, Grommet } from 'grommet';
+import { Route, Switch, useRouteMatch } from 'react-router-dom';
+import { AddBlogPost } from './AddBlogPost';
+import { BlogPostView } from './BlogPostView';
+import { BlogFeed } from './BlogFeed';
 
 // text editors
 // https://reactjsexample.com/tag/editor/
@@ -33,11 +36,7 @@ function Blog({user, error, loading }: WrappedComponentProps) {
                     <div className="col-12">
                         <Switch>
                             <Route exact path={path}>
-                                {(!posts || posts.length < 1) && <ShowLoading />}
-                                {posts.map((p, i) => 
-                                    <div className="pt-4" key={i}>
-                                        <ShortenedBlogPostView {...p} />
-                                    </div>)}
+                                <BlogFeed posts={posts} />
                             </Route>
                             <Route path={`${path}/:postId`}>
                                 <BlogPostView />
@@ -52,73 +51,6 @@ function Blog({user, error, loading }: WrappedComponentProps) {
     function refreshPosts(){
         if(user) blogPostRepo.readUserPosts(user.uid).then(r => setPosts(r));
     }
-}
-
-function AddBlogPost({ user, refreshPosts }: {user: firebase.User, refreshPosts: () => void}) {
-    let [content, setContent] = React.useState("");
-    let blogPostRepo = new BlogPostRepository();
-
-    return (
-        <div className="row">
-            <div className="col-12 d-flex flex-column">
-                <TextArea value={content} onChange={e => setContent(e.currentTarget.value)} />
-                <div className="pt-3 d-flex justify-content-end">
-                    <Button primary label="Submit" onClick={onAdd} />
-                </div>
-            </div>
-        </div>
-    );
-
-    function onAdd(e) {
-        blogPostRepo
-            .writeNewPost({ 
-                uid: user.uid, 
-                author: user.displayName, 
-                title: "testing", 
-                body: content 
-            })
-            .finally(refreshPosts);
-
-        setContent("");
-    }
-}
-
-function ShortenedBlogPostView({ primaryKey, post }: IBlogPostWithKey) {
-    let {title, body, author } = post;
-    let history = useHistory();
-
-    return (
-        <div style={{ display: "flex", flexDirection: "column" }}>
-            <h3>{title}</h3>
-            <Markdown>{body}</Markdown>
-            <div>
-                <Button label={"Continue Reading..."} onClick={() => {
-                    history.push(`blog/${primaryKey}`);
-                }} />
-            </div>
-            {/* <small>written by {author}</small> */}
-        </div>
-    );
-}
-
-function BlogPostView(props){
-    let { postId } = useParams();
-    let [post, setPost] = React.useState<BlogPost>(null);
-    
-    React.useEffect(() => {
-        new BlogPostRepository()
-            .readPost(postId)
-            .then(r => setPost(r.post));
-    });
-
-    if(!post) return <ShowLoading />;
-
-    return (
-        <div style={{ display: "flex", flexDirection: "column" }}>
-            <h3>{post.title}</h3>
-            <Markdown>{post.body}</Markdown>
-        </div>
-    );
 }
 
 export default createWithAuth(Blog);
